Correção de erros e importações do código original
