--- src/caixa.rs
@@ -1,6 +1,0 @@
-pub struct Caixa{
-
-
-
-    
-}
