--- src/carteiro.rs
@@ -1,16 +1,0 @@
-enum Status{
-    menu,
-    jogando_sem_caixa,
-    jogando_com_caixa,
-    fim
-}
-
-pub struct Carteiro{
-    pos_x: i32,
-    pos_y: i32,
-    status: Status
-}
-
-impl Carteiro{
-    
-}
