+++ src/components/caixa.rs
@@ -1,0 +1,14 @@
+pub struct Caixa{
+
+    
+}
+
+
+impl Caixa {
+    pub fn new() -> Self{
+        //aqui tem que retornar o objeto da Caixa
+
+        Caixa {}
+    }
+    
+}
