--- src/jogo.rs
@@ -1,18 +1,0 @@
-mod Carteiro;
-mod Caixa;
-
-
-pub struct Jogo{
-    carteiro: Carteiro,
-    caixa: Caixa
-}
-
-impl Jogo {
-    pub fn joga(&self){
-    
-    }
-
-    pub fn cria_jogo(matriz: [[char; 3] ; 3]){
-
-    }
-    }
